pyproject.toml
[tool.poetry]
name = "KoreAdmin"
version = "0.1.0"
description = ""
authors = ["kliuoi <kliuoi@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# 程序主体
fastapi = "^0.110.0"
# ORM
sqlalchemy = "^2.0.28"
# ORM异步Sqlite依赖
aiosqlite = "^0.20.0"
# ORM异步PostgreSQL依赖
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
# 环境变量配置
pydantic-settings = "^2.2.1"
# HTTP服务器
uvicorn = "^0.29.0"
# 模板引擎
jinja2 = "^3.1.3"
# 命令行
click = "^8.1.7"
# 密码加密
passlib = "^1.7.4"
# JWT
pyjwt = "^2.8.0"
# 时区依赖
tzdata = "^2024.1"
# 程序Session依赖
itsdangerous = "^2.1.2"
# fastapi表单依赖
python-multipart = "^0.0.9"
# Casbin
casbin = "^1.36.0"
casbin-async-sqlalchemy-adapter = "^1.4.0"
# 数据迁移工具
alembic = "^1.13.1"
# 异步支持
asgiref = "^3.8.1"
# 返回json格式化
orjson = "^3.10.6"
# 测试依赖
pytest = "^8.1.1"
httpx = "^0.27.0"
# 测试覆盖率
coverage = "^7.5.4"
# 测试异步支持
pytest-asyncio = "^0.23.7"
requests = "^2.32.3"







[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
